+++ xash3d_engine/engine/client/s_mouth.c
@@ -54,10 +54,10 @@ void SND_MoveMouth8(channel_t* ch, wavdata_t* pSource, int count)
 	if ( ch->isSentence )
 	{
 		if ( ch->currentWord )
-			pos = ch->currentWord->sample;
+			pos = (int)ch->currentWord->sample;
 	}
 	else
-		pos = ch->pMixer.sample;
+		pos = (int)ch->pMixer.sample;
 
 	count = S_GetOutputData(pSource, (void**)&pdata, pos, count, ch->use_loop);
 	if ( pdata == NULL )
@@ -81,7 +81,7 @@ void SND_MoveMouth8(channel_t* ch, wavdata_t* pSource, int count)
 
 	if ( pMouth->sndcount >= CAVGSAMPLES )
 	{
-		pMouth->mouthopen = pMouth->sndavg / CAVGSAMPLES;
+		pMouth->mouthopen = (byte)(pMouth->sndavg / CAVGSAMPLES);
 		pMouth->sndavg = 0;
 		pMouth->sndcount = 0;
 	}
@@ -105,10 +105,10 @@ void SND_MoveMouth16(channel_t* ch, wavdata_t* pSource, int count)
 	if ( ch->isSentence )
 	{
 		if ( ch->currentWord )
-			pos = ch->currentWord->sample;
+			pos = (int)ch->currentWord->sample;
 	}
 	else
-		pos = ch->pMixer.sample;
+		pos = (int)ch->pMixer.sample;
 
 	count = S_GetOutputData(pSource, (void**)&pdata, pos, count, ch->use_loop);
 	if ( pdata == NULL )
@@ -133,7 +133,7 @@ void SND_MoveMouth16(channel_t* ch, wavdata_t* pSource, int count)
 
 	if ( pMouth->sndcount >= CAVGSAMPLES )
 	{
-		pMouth->mouthopen = pMouth->sndavg / CAVGSAMPLES;
+		pMouth->mouthopen = (byte)(pMouth->sndavg / CAVGSAMPLES);
 		pMouth->sndavg = 0;
 		pMouth->sndcount = 0;
 	}
@@ -199,7 +199,7 @@ void SND_MoveMouthRaw(rawchan_t* ch, portable_samplepair_t* pData, int count)
 
 	if ( pMouth->sndcount >= CAVGSAMPLES )
 	{
-		pMouth->mouthopen = pMouth->sndavg / CAVGSAMPLES;
+		pMouth->mouthopen = (byte)(pMouth->sndavg / CAVGSAMPLES);
 		pMouth->sndavg = 0;
 		pMouth->sndcount = 0;
 	}
